refactor: Customer 관련 파일 패키지 위치 변경
