+++ src/test/java/com/programmers/springbootjpa/domain/customer/CustomerTest.java
@@ -1,5 +1,6 @@
-package com.programmers.springbootjpa.domain;
+package com.programmers.springbootjpa.domain.customer;
 
+import com.programmers.springbootjpa.domain.customer.Customer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
